moved all modules inside core, moved common into arcana-common package
