--- arcana/common/analysis/__init__.py
@@ -1,1 +1,0 @@
-# Import all Analysis classes in the namespace package here
