--- arcana/common/analysis/tasks/__init__.py
@@ -1,2 +1,0 @@
-from .archive import create_zip, create_tar, extract_zip, extract_tar
-from .shell import shell_cmd
