--- arcana/common/deploy.py
@@ -1,1 +1,0 @@
-from arcana.core.deploy.image.app import App  # noqa: F401
