+++ arcana/core/analysis/pipeline.py
@@ -8,7 +8,7 @@ from copy import copy, deepcopy
 import attrs.converters
 import pydra.mark
 from pydra.engine.core import Workflow
-from arcana.exceptions import (
+from arcana.core.exceptions import (
     ArcanaNameError,
     ArcanaUsageError,
     ArcanaDesignError,
