+++ arcana/core/analysis/spec.py
@@ -9,7 +9,7 @@ from attrs.converters import default_if_none
 from ..data.space import DataSpace
 from .salience import CheckSalience, ColumnSalience, ParameterSalience
 from ..utils.misc import ARCANA_SPEC
-from arcana.exceptions import ArcanaDesignError
+from arcana.core.exceptions import ArcanaDesignError
 
 
 @attrs.define
