+++ arcana/core/cli/base.py
@@ -2,12 +2,12 @@ from pathlib import Path
 import pkgutil
 from importlib import import_module
 import click
-import arcana
+import arcana.core as core
 
 
 # Define the base CLI entrypoint
 @click.group()
-@click.version_option(version=arcana.__version__)
+@click.version_option(version=core.__version__)
 def cli():
     """Base command line group, installed as "arcana"."""
 
@@ -17,7 +17,7 @@ def ext():
     """Command-line group for extension hooks"""
 
 
-CLI_EXT_PKG = Path(arcana.__file__).parent / "cli"
+CLI_EXT_PKG = Path(core.__file__).parent / "cli"
 
 # Ensure that all sub-packages under CLI are loaded so they are added to the
 # base command
