+++ arcana/core/cli/deploy.py
@@ -18,7 +18,7 @@ from arcana.core.utils.serialize import (
 from arcana.core.deploy.image import Metapackage, App
 
 # from arcana.deploy.xnat.image import XnatCSImage
-from arcana.exceptions import ArcanaBuildError
+from arcana.core.exceptions import ArcanaBuildError
 from arcana.core.data.set import Dataset
 from arcana.core.data.store import DataStore
 from arcana.core.utils.misc import extract_file_from_docker_image, DOCKER_HUB
