+++ arcana/core/data/column.py
@@ -7,7 +7,7 @@ from attrs.converters import optional
 
 # from arcana.core.data.row import DataRow
 from arcana.core.utils.serialize import ClassResolver
-from arcana.exceptions import ArcanaDataMatchError
+from arcana.core.exceptions import ArcanaDataMatchError
 from ..analysis.salience import DataQuality, ColumnSalience
 from .space import DataSpace
