+++ arcana/core/data/type/base.py
@@ -15,7 +15,7 @@ from arcana.core.utils.misc import (
     CONVERTER_ANNOTATIONS,
     HASH_CHUNK_SIZE,
 )
-from arcana.exceptions import (
+from arcana.core.exceptions import (
     ArcanaUsageError,
     ArcanaNameError,
     ArcanaDataNotDerivedYetError,
