+++ arcana/core/data/type/file.py
@@ -8,7 +8,7 @@ import logging
 import shutil
 import attrs
 from attrs.converters import optional
-from arcana.exceptions import (
+from arcana.core.exceptions import (
     ArcanaUsageError,
     ArcanaFileFormatError,
 )
