+++ arcana/core/deploy/command/components.py
@@ -11,7 +11,7 @@ from arcana.core.analysis.pipeline import (
 )
 from arcana.core.data.row import DataRow
 from arcana.core.data.type.base import DataType
-from arcana.exceptions import ArcanaFormatConversionError
+from arcana.core.exceptions import ArcanaFormatConversionError
 from arcana.core.data.space import DataSpace
 from arcana.core.utils.misc import add_exc_note
