+++ arcana/core/exceptions.py
